Refactor pixel worker to use reusable client module
